% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{add_scenario}
\alias{add_scenario}
\title{Add a scenario to a dsc.}
\usage{
add_scenario(dsc, name, fn, args = NULL, seed, metatype = "default_meta",
  inputtype = "default_input")
}
\arguments{
\item{dsc}{The dsc to add the sceanario to.}

\item{name}{A character string.}

\item{fn}{The datamaker, a function that is the datamaker for the
scenario.}

\item{args}{A list of arguments to the datamaker.}

\item{seed}{A vector of integers showing seeds to use for the scenario.}
}
\value{
Nothing, but modifies the dsc environment.
}
\description{
Adds a scenario to a dsc.
}
