% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{addFilenames}
\alias{addFilenames}
\title{Add outputfilenames to the database}
\usage{
addFilenames(dsc, outputfiletype)
}
\arguments{
\item{dsc}{the dsc to use}
}
\value{
nothing, but modified dsc
}
\description{
Add outputfilenames
}

