% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{addMethod}
\alias{addMethod}
\title{Deprecated alias to \code{\link{add_method}}.}
\usage{
addMethod(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{add_method}}.}
}
\description{
Deprecated alias to \code{\link{add_method}}.
}
