% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{addOutputParser}
\alias{addOutputParser}
\title{Add a outputParser to a dsc}
\usage{
addOutputParser(dsc, name, fn, outputtype_from = "default_output",
  outputtype_to)
}
\arguments{
\item{dsc}{the dsc to add the outputParser}

\item{name}{string giving a name by which the outputParser function is to be known by}

\item{fn,}{a outputParser function}

\item{outputtype_from,}{string naming the type of output the outputParser function takes as input}

\item{outputtype_to,}{string naming the type of output the outputParser function gives as output}
}
\value{
nothing, but creates output files in output/outputtype_to subdir
}
\description{
Adds a outputParser to a dsc; a outputParser converts one type of output to another type
}

