% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{addParser}
\alias{addParser}
\title{Add a parser to a dsc}
\usage{
addParser(dsc, name, fn, outputtype, methodnames)
}
\arguments{
\item{dsc}{the dsc to add the parser}

\item{name}{string giving a name by which the parser function is to be known by}

\item{fn,}{a parser function}

\item{outputtype,}{string naming the type of output the parser function creates.}

\item{methodnames,}{a list of methodnames that the parser can be applied to}
}
\value{
nothing, but modifies the dsc environment
}
\description{
Adds a parser to a dsc; a parser converts one type of output to another type
}

