% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{addScenario}
\alias{addScenario}
\title{Deprecated alias to \code{\link{add_scenario}}.}
\usage{
addScenario(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{add_scenario}}.}
}
\description{
Deprecated alias to \code{\link{add_scenario}}.
}
