% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/main.R
\name{addScenarioGroup}
\alias{addScenarioGroup}
\title{Add a set of scenarios to a dsc}
\usage{
addScenarioGroup(dsc, args, flexible.args,
  flexible.args.name = names(flexible.args), name = NULL, ...)
}
\arguments{
\item{args}{Arguments that should be passed to
all simulations}

\item{flexible.args}{data.frame of argument combinations;
    each combinations corresponds to one scenario, the names
of the argument are taken from the names() attribute}

\item{flexible.args.name}{str of the names of the columns
in flexible.args, should have the same length}

\item{name}{function or character or NULL
 if function, this will get passed a set of flexible
args and should return a scenario name
if NULL, a default title is generated
if of type character, parameters are added with
underscore separaters}

\item{...}{further arguments passed directly to addScenario}
}
\description{
Adds a series of simulations with the same call structure
as scenarios.
}
\examples{
\dontrun{
dsc = new.dsc("example_dsc", "example_dsc")
datamaker <- function(args){
output <- list()
output$input <- rnorm(100,args$mean, args$sd)
output$meta <- c()
}
addScenarioGroup(dsc, list(), expand.grid(mean=-2:2, sd=1:3),
name="norm", fn=datamaker, seed=1:5)
}
}

