% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{addScore}
\alias{addScore}
\title{Deprecated alias to \code{\link{add_score}}.}
\usage{
addScore(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{add_score}}.}
}
\description{
Deprecated alias to \code{\link{add_score}}.
}
