% Generated by roxygen2 (4.0.2): do not edit by hand
\name{aggregate_results}
\alias{aggregate_results}
\title{Aggregate the results of multiple methods for multiple trials}
\usage{
aggregate_results(scenarios, methods)
}
\arguments{
\item{scenarios}{a list of scenarios.}

\item{methods}{a list of methods}
}
\value{
a data frame of results, with one row for each trial/method combination. The details of the columns will depend on the comparison being run
}
\description{
Aggregate the results of multiple methods for multiple trials
}

