% Generated by roxygen2 (4.0.1): do not edit by hand
\name{apply_method}
\alias{apply_method}
\title{Apply a method to all inputs and produce corresponding outputs}
\usage{
apply_method(scenario_seedlist, method, flavor = NULL)
}
\arguments{
\item{scenario_seedlist}{(list), one element for each scenario, each element contains the vector of seeds for the pseudo-rng which identifies/indexes trials}

\item{method}{a list with elements
\item{methodname}{string by which method should be identified}
\item{methodfn}{name of function that is used to call the method}}
}
\value{
none - the results are saved in a file determined by outputfile(seed, method$methodname)
}
\description{
Apply a method to all inputs and produce corresponding outputs (by repeated application of apply_method_once).
Results are saved in a file
}

