% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{apply_method_..m}
\alias{apply_method_..m}
\title{Apply method to all scenarios}
\usage{
apply_method_..m(scenarios, method)
}
\arguments{
\item{scenarios}{a list of scenarios}

\item{method}{a list containing the method name, fn, etc}
}
\value{
none; data are saved in files in the output subdirectory
}
\description{
Apply method to all scenarios
}

