% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{apply_method_.s.}
\alias{apply_method_.s.}
\title{Apply all methods to a scenario}
\usage{
apply_method_.s.(scenario, methods)
}
\arguments{
\item{scenario}{a scenario}

\item{methods}{a list of methods}
}
\value{
none; data are saved in files in the output subdirectory
}
\description{
Apply all methods to a scenario
}

