% Generated by roxygen2 (4.0.2): do not edit by hand
\name{apply_method_.sm}
\alias{apply_method_.sm}
\title{Apply a method to all trials for a particular scenario}
\usage{
apply_method_.sm(scenario, method)
}
\arguments{
\item{scenario}{a list including elements seed and name}

\item{method}{a list including fn and name}
}
\value{
none; output are saved in the output subdirectory
}
\description{
Apply a method to all trials for a particular scenario
}

