% Generated by roxygen2 (4.0.2): do not edit by hand
\name{apply_method_rsm}
\alias{apply_method_rsm}
\title{Apply a method for a particular seed and scenario}
\usage{
apply_method_rsm(seed, scenario, method)
}
\arguments{
\item{seed}{(vector or list of integers) the seeds for the pseudo-rng which identifies/indexes trials}

\item{scenario}{a list including elements fn and args, the function name for the datamaker to be used and additional arguments}

\item{method}{a list including elements fn, name and args}
}
\value{
none; output are saved in the output subdirectory
}
\description{
Apply a method for a particular seed and scenario. THe subscript _rsm refers to a particular repetition
of a particular scenario for particular method.
}

