% Generated by roxygen2 (4.0.1): do not edit by hand
\name{apply_method_singletrial}
\alias{apply_method_singletrial}
\title{Apply a method to an input and produce output}
\usage{
apply_method_singletrial(indexlist, method, flavor = NULL)
}
\arguments{
\item{indexlist}{list that ``indexes" the comparison being used, with components
\item{seed}{the seed for the pseudo-rng which identifies/indexes the file}
\item{scenario}{string indicating scenario name}}

\item{method}{a list with elements
\item{methodname}{string by which method should be identified}
\item{methodfn}{name of function that is used to call the method}}

\item{flavor}{a string indicating which element of methods$flavorlist to use as additional arguments}
}
\value{
output a list of appropriate format to be determined by the comparison being run
}
\description{
Apply a single method to a single input trial and produce (and save) corresponding output
}

