% Generated by roxygen2 (4.0.2): do not edit by hand
\name{apply_methods}
\alias{apply_methods}
\title{Apply all methods to all scenarios}
\usage{
apply_methods(scenarios, methods)
}
\arguments{
\item{scenarios}{a list of scenarios}

\item{methods}{a list of methods}
}
\value{
none; data are saved in files in the output subdirectory
}
\description{
Apply all methods to all scenarios
}

