% Generated by roxygen2 (4.0.1): do not edit by hand
\name{datafile}
\alias{datafile}
\title{return the path to a data file, parameter file, output file or results file}
\usage{
datafile(indexlist, datadir = "data")
}
\arguments{
\item{indexlist}{list that ``indexes" the comparison being used, with components
\itemize{
\item{"seed"}{the seed for the pseudo-rng which identifies/indexes the file}
\item{"scenario"}{string indicating scenario name}
}}

\item{datadir/paramdir/outputdir/resultsdir}{the (relative) path to the directory containing the relevant files}
}
\value{
string containing path to file
}
\description{
return the path to a data file, parameter file, output file or results file
}

