% Generated by roxygen2 (4.0.2): do not edit by hand
\name{datafilename}
\alias{datafilename}
\title{return the path to a data file, parameter file, output file or results file}
\usage{
datafilename(seed, scenario, datadir = "data")
}
\arguments{
\item{seed}{}

\item{scenario}{}

\item{datadir/paramdir/outputdir/resultsdir}{the (relative) path to the directory containing the relevant files}
}
\value{
string containing path to file
}
\description{
return the path to a data file, parameter file, output file or results file
}

