% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dsc2BE}
\alias{dsc2BE}
\title{Create a BatchExperiments registry from a dsc}
\usage{
dsc2BE(dsc, id, repl = 100)
}
\arguments{
\item{dsc}{a list with components scenarios and methods and scorefn}

\item{id}{a string for the id}
}
\value{
a registry reg for BatchExperiments
}
\description{
Create a BatchExperiments registry from a dsc
}
\details{

}

