% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dsc_to_be.R
\name{dsc_to_be}
\alias{dsc_to_be}
\title{Create a BatchExperiments registry from a dsc}
\usage{
dsc_to_be(dsc, id, repl = 100)
}
\arguments{
\item{dsc}{a list with components scenarios and methods and scorefn}

\item{id}{a string for the id}
}
\value{
a registry reg for BatchExperiments
}
\description{
Create a BatchExperiments registry from a dsc
}
\details{

}

