% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\docType{package}
\name{dscr-package}
\alias{dscr}
\alias{dscr-package}
\title{Dynamic Statistical Comparisons in R}
\description{
Package containing functions to help build dynamic statistical comparisons in R

\tabular{ll}{ Package: \tab dscr\cr Type: \tab Package\cr
Version: \tab 0.1\cr License: \tab GPL (>=2)\cr }
}
\author{
Matthew Stephens <\email{mstephens@uchicago.edu}>
}
\keyword{dscr}

