% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{dscr}
\alias{dscr}
\alias{dscr-package}
\title{Perform Dynamic Statistical Comparisons in R}
\description{
dscr provides some functions for performing Dynamic Statistical Comparisons in R
}
\details{
Run browseVignettes("dscr") for an example
}

