% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/expand_and_filter.R
\name{expand_dsc}
\alias{expand_dsc}
\title{Comprehensively tabulate a phase of \code{dscr} execution.}
\usage{
expand_dsc(dsc, phase)
}
\arguments{
\item{dsc}{The \code{dscr} data structure to expand.}

\item{phase}{The phase of execution to tabulate, either
  'scenarios', 'scenarios_methods', and 'scenarios_methods_scores'.}
}
\value{
Data frame with columns \code{scenarioname}, \code{seed},
  \code{methodname} (if one of the latter two phases is requested),
  and \code{scorename} (if the last phase is selected).
}
\description{
Depending of the phase of execution requested, compute the
Cartesian product of all scenario-seed combinations (as specified
in \code{dsc$scenarios}), the set of methods (as specified in
\code{dsc$methods}), and the set of scores (as specified in
\code{dsc$scores}).
}

