% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{expand_scenario}
\alias{expand_scenario}
\title{Create a dataframe of scenarioname and seed combinations}
\usage{
expand_scenario(scenario)
}
\arguments{
\item{scenario}{the scenario to be expanded}
}
\value{
data frame of scenarioname and seed combinations
}
\description{
Create a dataframe of scenarioname and seed combinations
}

