% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{expand_scenarios}
\alias{expand_scenarios}
\title{Create a dataframe of all scenarioname and seed combinations}
\usage{
expand_scenarios(dsc)
}
\arguments{
\item{dsc}{the dsc to be expanded}
}
\value{
data frame of all scenarioname and seed combinations
}
\description{
Create a dataframe of all scenarioname and seed combinations
}

