% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{expand_scenarios_methods}
\alias{expand_scenarios_methods}
\title{Create a list of scenarioname, seed and method combinations}
\usage{
expand_scenarios_methods(dsc)
}
\arguments{
\item{dsc}{the dsc to expand}
}
\value{
data frame of all combinations
}
\description{
Create a list of scenarioname, seed and method combinations
}

