% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{expand_scenarios_methods_scores}
\alias{expand_scenarios_methods_scores}
\title{Create a dataframe of scenarioname, seed, method and score combinations}
\usage{
expand_scenarios_methods_scores(dsc)
}
\arguments{
\item{dsc}{the dsc to expand}
}
\value{
data frame of all combinations
}
\description{
Create a dataframe of scenarioname, seed, method and score combinations
}

