% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{expandScenario}
\alias{expandScenario}
\title{Create a dataframe of scenarioname and seed combinations}
\usage{
expandScenario(scenario)
}
\arguments{
\item{scenario}{the scenario to be expanded}
}
\value{
data frame of scenarioname and seed combinations
}
\description{
Create a dataframe of scenarioname and seed combinations
}

