% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/main.R
\name{expandScenarios}
\alias{expandScenarios}
\title{Create a dataframe of all scenarioname and seed combinations}
\usage{
expandScenarios(dsc)
}
\arguments{
\item{dsc}{the dsc to be expanded}
}
\value{
data frame of all scenarioname and seed combinations
}
\description{
Create a dataframe of all scenarioname and seed combinations
}

