% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{expandScenariosMethods}
\alias{expandScenariosMethods}
\title{Create a list of scenarioname, seed and method combinations}
\usage{
expandScenariosMethods(dsc)
}
\arguments{
\item{dsc}{the dsc to expand}
}
\value{
data frame of all combinations
}
\description{
Create a list of scenarioname, seed and method combinations
}

