% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/main.R
\name{expandScenariosMethodsScores}
\alias{expandScenariosMethodsScores}
\title{Create a dataframe of scenarioname, seed, method and score combinations}
\usage{
expandScenariosMethodsScores(dsc)
}
\arguments{
\item{dsc}{the dsc to expand}
}
\value{
data frame of all combinations
}
\description{
Create a dataframe of scenarioname, seed, method and score combinations
}

