% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsc.R
\name{get_meta_types}
\alias{get_meta_types}
\title{Return metatypes of the scenarios in a dsc.}
\usage{
get_meta_types(dsc)
}
\arguments{
\item{dsc}{A dsc.}
}
\value{
List of inputtypes.
}
\description{
Return metatypes of the scenarios in a dsc.
}
