% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_one_result}
\alias{get_one_result}
\title{Get the results of a single method for a single trial}
\usage{
get_one_result(indexlist, method, flavor = NULL)
}
\arguments{
\item{indexlist}{list that ``indexes" the comparison being used, with components
\item{seed}{the seed for the pseudo-rng which identifies/indexes the file}
\item{scenario}{string indicating scenario name}}

\item{method}{a list with elements
\item{methodname}{string by which method should be identified}
\item{methodfn}{name of function that is used to call the method}}
}
\value{
results, a data frame of results, the details will depend on the comparison being run
}
\description{
Get the results of a single method for a single trial
}

