% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_output_types}
\alias{get_output_types}
\title{return outputtypes of the methods in a dsc}
\usage{
get_output_types(dsc)
}
\arguments{
\item{dsc}{a dsc}
}
\value{
list of outputtypes
}
\description{
return outputtypes of the methods in a dsc
}
