% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_results}
\alias{get_results}
\title{Get the results of a single method for multiple trials (one flavor)}
\usage{
get_results(method, scenario_seedlist, flavor = NULL)
}
\arguments{
\item{method}{a list with elements
\item{methodname}{string by which method should be identified}
\item{methodfn}{name of function that is used to call the method}}

\item{seed}{(list of integers) the seeds for the pseudo-rng which identifies/indexes trials}
}
\value{
a data frame of results, with one row for each trial. The details of the columns will depend on the comparison being run
}
\description{
Get the results of a single method for multiple trials (one flavor)
}

