% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_results_singletrial}
\alias{get_results_singletrial}
\title{Get the results of a single method for a single trial}
\usage{
get_results_singletrial(seed, scenario, method)
}
\arguments{
\item{seed}{}

\item{scenario}{}

\item{method}{}
}
\value{
results, a data frame of results, the details will depend on the comparison being run
}
\description{
Get the results of a single method for a single trial
}

