% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{getDatatypes}
\alias{getDatatypes}
\title{return datatypes of the scenarios in a dsc}
\usage{
getDatatypes(dsc)
}
\arguments{
\item{dsc}{a dsc}
}
\value{
list of datatypes
}
\description{
return datatypes of the scenarios in a dsc
}

