% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{getOutputtypes}
\alias{getOutputtypes}
\title{Deprecated alias to \code{\link{get_output_types}}.}
\usage{
getOutputtypes(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{get_output_types}}.}
}
\description{
Deprecated alias to \code{\link{get_output_types}}.
}
