% Generated by roxygen2 (4.0.2): do not edit by hand
\name{inspect_data_singletrial}
\alias{inspect_data_singletrial}
\title{provide the data of a single scenario for a single trial}
\usage{
inspect_data_singletrial(seed, scenario)
}
\arguments{
\item{seed}{}

\item{scenario}{}
}
\value{
output from method, the details will depend on the comparison being run
}
\description{
provide the data of a single scenario for a single trial; intended primarily for troubleshooting and debugging
}

