% Generated by roxygen2 (4.0.2): do not edit by hand
\name{inspect_results_singletrial}
\alias{inspect_results_singletrial}
\title{provide the results of a single method for a single trial}
\usage{
inspect_results_singletrial(seed, scenario, method)
}
\arguments{
\item{seed}{}

\item{scenario}{}

\item{method}{}
}
\value{
results output by score function, the details will depend on the comparison being run
}
\description{
provide the results of a single method for a single trial; intended primarily for troubleshooting and debugging
}

