% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{list_scenarios}
\alias{list_scenarios}
\title{List all scenarios for a DSC}
\usage{
list_scenarios(dsc)
}
\arguments{
\item{scenarios}{a list of scenarios used to produce data=list(input,meta)}
}
\value{
a list of names of the scenarios
}
\description{
List all scenarios for a DSC
}

