% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{listoutputParsers}
\alias{listoutputParsers}
\title{Deprecated alias to \code{\link{list_output_parsers}}.}
\usage{
listoutputParsers(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{list_output_parsers}}.}
}
\description{
Deprecated alias to \code{\link{list_output_parsers}}.
}
