% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/main.R
\name{listScenarios}
\alias{listScenarios}
\title{List scenarios}
\usage{
listScenarios(dsc)
}
\arguments{
\item{dsc}{the dsc to have its scenarios listed}
}
\value{
nothing
}
\description{
List scenarios
}

