% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsc.R
\name{load_example}
\alias{load_example}
\title{Return the data and output for a single method for a single
  trial.}
\usage{
load_example(dsc, seed, scenarioname, methodname, homedir = ".")
}
\arguments{
\item{dsc}{A dsc object.}

\item{scenarioname}{Name of the  selected scenario.}

\item{methodname}{Name of the selected method.}

\item{homedir}{The directory from which the dsc was run.}
}
\value{
Results list with components \code{data} and \code{output}.
}
\description{
Return a list containing data and output for a single
  trial.
}
