% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{loadExample}
\alias{loadExample}
\title{Deprecated alias to \code{\link{load_example}}.}
\usage{
loadExample(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{load_example}}.}
}
\description{
Deprecated alias to \code{\link{load_example}}.
}
