% Generated by roxygen2 (4.0.2): do not edit by hand
\name{long_methodname}
\alias{long_methodname}
\title{combine a method name and flavor to produce a new method name}
\usage{
long_methodname(methodname, flavor = NULL)
}
\arguments{
\item{methodname}{(string) name of a method}

\item{flavor}{(sting) name of a flavor}
}
\value{
string method.flavor, or if flavor is null then just method
}
\description{
combine a method name and flavor to produce a new method name
}

