% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{make_data_rs}
\alias{make_data_rs}
\title{Make the data (inputs and meta) for a DSC for a particular seed and scenario}
\usage{
make_data_rs(seed, scenario, overwrite = FALSE)
}
\arguments{
\item{seed}{(vector or list of integers) the seeds for the pseudo-rng which identifies/indexes trials}

\item{scenario}{a list including elements fn and args, the function name for the datamaker to be used and additional arguments}

\item{overwrite}{boolean indicating whether to overwrite existing files; default is not to}
}
\value{
data are saved in files in the data subdirectory
}
\description{
Make the data (inputs and meta) for a DSC for a particular seed and scenario. THe subscript rs refers to the "repetition" of a "scenario".
I'm experiementing with naming conventions to try to make these easier to see, using . as a subscript to indicate
that that subscript is applied over. By this logic make_data_scenario should be make_data_.s
}

