% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{make_nice_score_names}
\alias{make_nice_score_names}
\title{Score a method on a single trial and save results}
\usage{
make_nice_score_names(s)
}
\arguments{
\item{score,}{the vector of scores with some elements possibly named}
}
\value{
a vector of names for a score. For elements of score already named, the name stays the same; for unnamed elements the name is scorej where j is the index of the element
}
\description{
tries to deal with scores that don't have names
}

