% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make_params}
\alias{make_params}
\title{Make all the parameters for a DSC}
\usage{
make_params(parammaker, scenario_seedlist)
}
\arguments{
\item{parammaker}{a function for making parameters from seeds and scenario combinations}

\item{scenario_seedlist}{a list of integer vectors, one for each scenario. Each list gives the seeds to be used for that scenario.}
}
\value{
parameters are saved in files in the params subdirectory
}
\description{
Make all the parameters for DSC using all combinations of seed and scenario
}

