% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{meta_file_name}
\alias{meta_file_name}
\title{return the path to a meta file, parameter file, output file or results file}
\usage{
meta_file_name(dsc, seed, scenario, metadir = "meta", metatype = NULL)
}
\arguments{
\item{scenario}{}

\item{inputdir/metadir/outputdir/scoresdir}{the (relative) path to the directory containing the relevant files}
}
\value{
string containing path to file
}
\description{
return the path to a meta file, parameter file, output file or results file
}
