% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{metafilename}
\alias{metafilename}
\title{return the path to a meta file, parameter file, output file or results file}
\usage{
metafilename(dsc, seed, scenario, metadir = "meta", metatype = NULL)
}
\arguments{
\item{dsc}{}

\item{seed}{}

\item{scenario}{}

\item{inputdir/metadir/outputdir/scoresdir}{the (relative) path to the directory containing the relevant files}
}
\value{
string containing path to file
}
\description{
return the path to a meta file, parameter file, output file or results file
}

