% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/dsc2BE.R
\name{msfn2afn}
\alias{msfn2afn}
\title{convert methodfn and scorefn to a problem function}
\usage{
msfn2afn(methodfn, scorefn, ...)
}
\description{
convert a method function (with input (input, args) to a algo function, with input dynamic)
Not needed, so commented out!
}

