% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dsc_to_be.R
\name{msfn_to_afn}
\alias{msfn_to_afn}
\title{convert methodfn and scorefn to a problem function}
\usage{
msfn_to_afn(methodfn, scorefn, ...)
}
\description{
convert a method function (with input (input, args) to a algo function, with input dynamic)
Not needed, so commented out!
}

