% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expand_and_filter.R
\name{multiple_filter}
\alias{multiple_filter}
\title{Apply multiple filters to a data frame.}
\usage{
multiple_filter(df, ...)
}
\arguments{
\item{df}{Data frame to filter.}

\item{...}{Set of filters to apply to the data frame. The names of
the list elements correspond to column names in \code{df},
while the values of the list elements correspond to the values
that the column will be restricted to.}
}
\description{
Build and execute a \code{dplyr::filter} call against a data frame
with an arbitrary number of filtering clauses, some of which can be
inactive by being set to \code{NULL}. For example,
\code{multiple_filter(df, some_col = c(1, 2), other_col = NULL)}
returns the value of \code{dplyr::filter(df, some_col \%in\% c(1,
2))}. Return the data frame unchanged if there are no active
filters.
}

