% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{new.dsc}
\alias{new.dsc}
\title{Deprecated alias to \code{\link{new_dsc}}.}
\usage{
new.dsc(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{new_dsc}}.}
}
\description{
Deprecated alias to \code{\link{new_dsc}}.
}
