% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsc.R
\name{new_dsc}
\alias{new_dsc}
\title{Start a new (empty) dsc.}
\usage{
new_dsc(name, file.dir)
}
\arguments{
\item{name}{A string containing the name for your dsc.}
}
\value{
dsc, an environment.
}
\description{
Returns an environment that is a dsc.
}
