% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/main.R
\name{reset_scenario}
\alias{reset_scenario}
\title{Removes all output and results for a scenario}
\usage{
reset_scenario(dsc, scenarioname, force = FALSE)
}
\arguments{
\item{scenarioname}{string indicating name of scenario to remove}

\item{force}{boolean, indicates whether to proceed without prompting user (prompt is to be implemented)}
}
\value{
nothing; simply delets files
}
\description{
Removes all output and results for a scenario; primary intended purpose is to force re-running of that scenario.
Works only for unix look-alikes?
}

