% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{run_output_parser}
\alias{run_output_parser}
\title{Run a outputParser in a dsc}
\usage{
run_output_parser(dsc, outputParsername)
}
\arguments{
\item{dsc}{the dsc to use}

\item{outputParsername}{string giving the name of the outputParser to be run}
}
\value{
nothing, but outputs files to output/ directories
}
\description{
Run the named outputParser to convert one type of output to another type.
The outputParser is run on all outputs in the appropriate output directory (output/outputtype_from)
where outputtype_from is defined when the outputParser is added to the dsc
}

