% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_output_parsers}
\alias{run_output_parsers}
\title{Run all outputParsers in a dsc}
\usage{
run_output_parsers(dsc)
}
\arguments{
\item{dsc}{the dsc to use}
}
\value{
nothing, but outputs files to output/ directories
}
\description{
Run all outputParsers to convert one type of output to another type. 
Each outputParser is run on all outputs in the appropriate output directory (output/outputtype_from)
where outputtype_from is defined when the outputParser is added to the dsc
}
