% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{runScore}
\alias{runScore}
\title{Score a method on a single trial and save results}
\usage{
runScore(dsc, seed, scenarioname, methodname, scorename)
}
\arguments{
\item{seed}{the seed to score}

\item{scenarioname}{the scenario to score}

\item{methodname}{the method to score}

\item{scorename}{the score to use}
}
\value{
results, a list of appropriate format to be determined by the comparison being run (maybe required to be a dataframe?)
}
\description{
Score results of a single method for a single trial and produce (and save) corresponding results
}

