% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{runOutputParsers}
\alias{runOutputParsers}
\title{Run all outputParsers in a dsc}
\usage{
runOutputParsers(dsc)
}
\arguments{
\item{dsc}{the dsc to use}
}
\value{
nothing, but outputs files to output/ directories
}
\description{
Run all outputParsers to convert one type of output to another type.
Each outputParser is run on all outputs in the appropriate output directory (output/outputtype_from)
where outputtype_from is defined when the outputParser is added to the dsc
}

