% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{runParsers}
\alias{runParsers}
\title{Run all parsers in a dsc}
\usage{
runParsers(dsc)
}
\arguments{
\item{dsc}{the dsc to use}
}
\value{
nothing, but outputs files to output/ directories
}
\description{
Run all parsers to convert one type of output to another type.
Each parser is run on all outputs in the appropriate output directory (output/outputtype_from)
where outputtype_from is defined when the parser is added to the dsc
}

