% Generated by roxygen2 (4.0.2): do not edit by hand
\name{score_method}
\alias{score_method}
\title{Score a method on all trials and save results}
\usage{
score_method(scenario_seedlist, method, scorefn, flavor = NULL)
}
\arguments{
\item{scenario_seedlist}{(list) the seeds for the pseudo-rng for each scenario which identifies/indexes trials}

\item{method}{a list with elements
\itemize{
\item{methodname}{string by which method should be identified}
\item{methodfn}{name of function that is used to call the method}
}}

\item{scorefn}{a function that scores output based on comparisons with input, parameters and metadata}
}
\value{
results, a list of appropriate format to be determined by the comparison being run (maybe required to be a dataframe?)
}
\description{
Score a method on all trials and save results (by repeated application of score_method_singletrial)
}

