% Generated by roxygen2 (4.0.1): do not edit by hand
\name{score_method_singletrial}
\alias{score_method_singletrial}
\title{Score a method on a single trial and save results}
\usage{
score_method_singletrial(indexlist, method, scorefn, flavor = NULL)
}
\arguments{
\item{indexlist}{list that ``indexes" the comparison being used, with components
\item{seed}{the seed for the pseudo-rng which identifies/indexes the file}
\item{scenario}{string indicating scenario name}}

\item{method}{a list with elements
\item{methodname}{string by which method should be identified}
\item{methodfn}{name of function that is used to call the method}}

\item{scorefn}{a function that scores output based on comparisons with input, parameters and metadata}
}
\value{
results, a list of appropriate format to be determined by the comparison being run (maybe required to be a dataframe?)
}
\description{
Score results of a single method for a single trial and produce (and save) corresponding results
}

