% Generated by roxygen2 (4.0.2): do not edit by hand
\name{score_methods}
\alias{score_methods}
\title{Score specified methods on specified scenarios}
\usage{
score_methods(scenarios, methods, scorefn)
}
\arguments{
\item{scenarios}{the scenarios to score}

\item{methods}{the methods to score}

\item{scorefn}{a function that scores output based on comparisons with input, parameters and metadata}
}
\value{
results, a list of appropriate format to be determined by the comparison being run (maybe required to be a dataframe?)
}
\description{
Score specified methods on specified scenarios
}

