% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/execution.R
\name{seeded_function_call}
\alias{seeded_function_call}
\title{Perform a seeded function call.}
\usage{
seeded_function_call(fxn_to_call, seed, mem_params, disk_params = NULL)
}
\arguments{
\item{fxn_to_call}{Function to execute.}

\item{seed}{Integer random seed, passed transparently to
\code{set.seed}.}

\item{mem_params}{Named parameter list of objects loaded at
invocation time to pass to function.}

\item{disk_params}{Optional data frame of input objects to read
  from disk. \code{disk_params$file_names} should give the paths to
  the input files. \code{disk_params$variable_names} should give the
  (character) names of the variable names expected by
  \code{fxn_to_call}, which will be parsed to symbols, bound to the
  loaded objects, and merged with \code{mem_params}.}
}
\value{
Return value of \code{fxn_to_call}.
}
\description{
Set a random seed and call a given function with a named parameter
list, part of which can be reconstituted from serialized RDS
objects.
}
\details{
This is meant to be the abstract unit of execution that can be
passed to a \code{batchMap} call to the \code{BatchJobs} package
but that is flexible enough to handle the heterogeneous needs of a
phase of \code{dscr}.
}

