% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dsc2BE.R
\name{sfn2pfn}
\alias{sfn2pfn}
\title{convert datamaker to problem fn}
\usage{
sfn2pfn(fn, ...)
}
\description{
convert a scenario (datamaker) function with input (args) to a problem function, with input (static)
}

