% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dsc_to_be.R
\name{sfn_to_pfn}
\alias{sfn_to_pfn}
\title{convert datamaker to problem fn}
\usage{
sfn_to_pfn(fn, ...)
}
\description{
convert a scenario (datamaker) function with input (args) to a problem function, with input (static)
}

