% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{shiny_plot}
\alias{shiny_plot}
\title{Plot results for DSC.}
\usage{
shiny_plot(res, s = "scenario", m = "method")
}
\arguments{
\item{res}{Results of a DSC.}
}
\value{
A shiny plot.
}
\description{
Interactive plot for results of DSC.
}
