% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsc.R
\name{source_dir}
\alias{source_dir}
\title{Sources all R files in a directory.}
\usage{
source_dir(path, trace = TRUE, ...)
}
\arguments{
\item{path}{The directory you want to source.}

\item{trace}{Whether to print the names of files being sourced.}

\item{...}{Additional arguments passed to \code{\link{source}}.}
}
\value{
No return value.
}
\description{
Sources all R files in a directory.
}
