% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sourceDir}
\alias{sourceDir}
\title{Sources all R files in a directory}
\usage{
sourceDir(path, trace = TRUE, ...)
}
\arguments{
\item{path}{the directory you want to source}

\item{trace}{whether to print the names of files being sourced}
}
\value{
none;
}
\description{
Sources all R files in a directory (comes from examples in ??source)
}

