% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{update_makefile}
\alias{update_makefile}
\title{Create or update a Makefile for the package}
\usage{
update_makefile(deps = NULL, run_methods = NULL, vignettes = NULL,
  output = NULL)
}
\arguments{
\item{deps}{depends of the package to check in Makefile}

\item{run_methods}{method names to run in Makefile}

\item{vignettes}{vignette names to built in Makefile}
}
\value{
Makefile for the package
}
\description{
Create or update a Makefile for the package.
}
\details{

}
\examples{
\donttest{
deps = c('devtools', 'roxygen2', 'knitr', 'plyr', 'psych', 'reshape2')
run_methods = c('one_sample_location')
vignettes = c('one_sample_location.rmd')
update_makefile(deps, run_methods, vignettes, output = '~/dscr/Makefile')}
}

