% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{addMethod}
\alias{addMethod}
\title{Add a method to a dsc}
\usage{
addMethod(dsc, name, fn, args = NULL, outputtype = "default_output")
}
\arguments{
\item{dsc}{the dsc to add the method to}

\item{name}{a character string name for the method}

\item{fn,}{a wrapper function that implements the method}

\item{args}{a list of additional arguments to fn}
}
\value{
nothing, but modifies the dsc environment
}
\description{
Adds a method to a dsc
}

