% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{add_score}
\alias{add_score}
\title{Add a score function to a dsc}
\usage{
add_score(dsc, fn, name = "default_score", outputtype = "default_output")
}
\arguments{
\item{dsc}{the dsc to add the score to}

\item{fn,}{a score function}

\item{name}{string giving a name by which the score function is to be known by}

\item{outputtype,}{the type of output the score function takes, as generated by a outputParser}
}
\value{
nothing, but modifies the dsc environment
}
\description{
Adds a score to a dsc
}

