% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{addOutputParser}
\alias{addOutputParser}
\title{Deprecated alias to \code{\link{add_output_parser}}.}
\usage{
addOutputParser(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{add_output_parser}}.}
}
\description{
Deprecated alias to \code{\link{add_output_parser}}.
}
