% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{addParser}
\alias{addParser}
\title{Add a parser to a dsc}
\usage{
addParser(dsc, name, fn, outputtype_from = "default_output", outputtype_to)
}
\arguments{
\item{dsc}{the dsc to add the parser}

\item{name}{string giving a name by which the parser function is to be known by}

\item{fn,}{a parser function}

\item{outputtype_from,}{string naming the type of output the parser function takes as input}

\item{outputtype_to,}{string naming the type of output the parser function gives as output}
}
\value{
nothing, but creates output files in output/outputtype_to subdir
}
\description{
Adds a parser to a dsc; a parser converts one type of output to another type
}

