% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/main.R
\name{addScenario}
\alias{addScenario}
\title{Add a scenario to a dsc}
\usage{
addScenario(dsc, name, fn, args = NULL, seed, metatype = "default_meta",
  inputtype = "default_input")
}
\arguments{
\item{dsc}{the dsc to add the sceanario to}

\item{name}{a character string}

\item{fn}{the datamaker, a function that is the datamaker for the scenario}

\item{args}{a list of arguments to the datamaker}

\item{seed}{a vector of integers showing seeds to use for the scenario}
}
\value{
nothing, but modifies the dsc environment
}
\description{
Adds a scenario to a dsc
}

