% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{addScore}
\alias{addScore}
\title{Add a score function to a dsc}
\usage{
addScore(dsc, name, fn, outputtype = "defaultoutput")
}
\arguments{
\item{dsc}{the dsc to add the score to}

\item{name}{string giving a name by which the score function is to be known by}

\item{fn,}{a score function}

\item{outputtype,}{the type of output the score function takes, as generated by a parser}
}
\value{
nothing, but modifies the dsc environment
}
\description{
Adds a score to a dsc
}

