% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{addScore}
\alias{addScore}
\title{Add a score function to a dsc}
\usage{
addScore(dsc, scorefn)
}
\arguments{
\item{dsc}{the dsc to add the score to}

\item{scorefn,}{a score function}
}
\value{
nothing, but modifies the dsc environment
}
\description{
Adds a score to a dsc
}

