% Generated by roxygen2 (4.0.1): do not edit by hand
\name{aggregate_results}
\alias{aggregate_results}
\title{Aggregate the results of multiple methods for multiple trials}
\usage{
aggregate_results(methodslist, scenario_seedlist)
}
\arguments{
\item{methodslist}{a list of methods. Each method is itself a list with elements
\item{methodname}{string by which method should be identified}
\item{methodfn}{name of function that is used to call the method}
\item{flavorlist}{list of flavors of the method (each flavor is a list containing additional parameters to be passed to methodfn)}}

\item{seed}{(list of integers) the seeds for the pseudo-rng which identifies/indexes trials}
}
\value{
a data frame of results, with one row for each trial/method combination. The details of the columns will depend on the comparison being run
}
\description{
Aggregate the results of multiple methods for multiple trials
}

