% Generated by roxygen2 (4.0.2): do not edit by hand
\name{apply_method_..m}
\alias{apply_method_..m}
\title{Apply method to all scenarios}
\usage{
apply_method_..m(scenarios, method)
}
\arguments{
\item{scenarios}{a list of scenarios}

\item{method}{a list containing the method name, fn, etc}
}
\value{
none; data are saved in files in the output subdirectory
}
\description{
Apply method to all scenarios
}

