% Generated by roxygen2 (4.0.2): do not edit by hand
\name{apply_method_singletrial}
\alias{apply_method_singletrial}
\title{Apply a method for a particular seed and scenario}
\usage{
apply_method_singletrial(seed, scenario, method)
}
\arguments{
\item{seed}{(vector or list of integers) the seeds for the pseudo-rng which identifies/indexes trials}

\item{scenario}{a list including elements fn and args, the function name for the datamaker to be used and additional arguments}
}
\value{
none; output are saved in the output subdirectory
}
\description{
Apply a method for a particular seed and scenario
}

