% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{datafilename}
\alias{datafilename}
\title{return the path to a data file, parameter file, output file or results file}
\usage{
datafilename(dsc, seed, scenario, datadir = "data")
}
\arguments{
\item{seed}{}

\item{scenario}{}

\item{datadir/paramdir/outputdir/scoresdir}{the (relative) path to the directory containing the relevant files}
}
\value{
string containing path to file
}
\description{
return the path to a data file, parameter file, output file or results file
}

