% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{package}
\name{dscr-package}
\alias{dscr-package}
\alias{dscr}
\title{Dynamic Statistical Comparisons in R}
\description{
Package containing functions to help build dynamic
  statistical comparisons in R.
}
\author{
Matthew Stephens \email{mstephens@uchicago.edu}
}
\keyword{dscr}
