% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/expand_and_filter.R
\name{expand_dsc}
\alias{expand_dsc}
\title{Comprehensively tabulate a phase of \code{dscr} execution.}
\usage{
expand_dsc(dsc, phase)
}
\arguments{
\item{dsc}{The \code{dscr} data structure to expand.}

\item{phase}{The phase of execution to tabulate, either
  'scenarios', 'scenarios_methods', or 'scenarios_methods_scores'.}
}
\value{
Data frame with columns \code{scenarioname}, \code{seed},
  \code{methodname} (if one of the latter two phases is requested),
  and \code{scorename} (if the last phase is selected).
}
\description{
Depending on the phase of execution requested, starting from
\enumerate{
  \item the set of scenario-seed combinations (as specified
    in \code{dsc$scenarios}),
  \item the set of methods (as specified in
    \code{dsc$methods}), and
  \item the set of scores (as specified in
    \code{dsc$scores}),
}
compute (1), the Cartesian product of (1) and (2), or the Cartesian
product of (1), (2), and (3).
}

