% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_and_filter.R
\name{expand_scenario}
\alias{expand_scenario}
\title{Tabulate a \code{dscr} scenario.}
\usage{
expand_scenario(scenario)
}
\arguments{
\item{scenario}{Scenario to be expanded, with \code{scenario$name}
giving a string for the scenario's name and \code{scenario$seed}
giving an integer vector of random seeds.}
}
\value{
Data frame with columns \code{scenarioname} and
  \code{seed}. The column \code{scenarioname} is coerced to a
  character vector.
}
\description{
Given a scenario, return a data frame tabulating a scenario's name
in one column and its set of seeds in another column.
}
