% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{getInputtypes}
\alias{getInputtypes}
\title{return inputtypes of the scenarios in a dsc}
\usage{
getInputtypes(dsc)
}
\arguments{
\item{dsc}{a dsc}
}
\value{
list of inputtypes
}
\description{
return inputtypes of the scenarios in a dsc
}

