% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_input_types}
\alias{get_input_types}
\title{return inputtypes of the scenarios in a dsc}
\usage{
get_input_types(dsc)
}
\arguments{
\item{dsc}{a dsc}
}
\value{
list of inputtypes
}
\description{
return inputtypes of the scenarios in a dsc
}
