% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{getMetatypes}
\alias{getMetatypes}
\title{return metatypes of the scenarios in a dsc}
\usage{
getMetatypes(dsc)
}
\arguments{
\item{dsc}{a dsc}
}
\value{
list of inputtypes
}
\description{
return metatypes of the scenarios in a dsc
}

