% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{get_meta_types}
\alias{get_meta_types}
\title{return metatypes of the scenarios in a dsc}
\usage{
get_meta_types(dsc)
}
\arguments{
\item{dsc}{a dsc}
}
\value{
list of inputtypes
}
\description{
return metatypes of the scenarios in a dsc
}

