% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_results_scenario}
\alias{get_results_scenario}
\title{Get the results of a single method for a single scenario}
\usage{
get_results_scenario(dsc, scenario, method, score)
}
\arguments{
\item{scenario}{a scenario}

\item{method}{a method}
}
\value{
a data frame of results, with one row for each trial. The details of the columns will depend on the comparison being run
}
\description{
Get the results of a single method for a single scenario
}
