% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/main.R
\name{getOutputtypes}
\alias{getOutputtypes}
\title{return outputtypes of the methods in a dsc}
\usage{
getOutputtypes(dsc)
}
\arguments{
\item{dsc}{a dsc}
}
\value{
list of outputtypes
}
\description{
return outputtypes of the methods in a dsc
}

