% Generated by roxygen2 (4.0.2): do not edit by hand
\name{inspect_output_singletrial}
\alias{inspect_output_singletrial}
\title{provide the output of a single method for a single trial}
\usage{
inspect_output_singletrial(seed, scenario, method)
}
\arguments{
\item{seed}{}

\item{scenario}{}

\item{method}{}
}
\value{
output from method, the details will depend on the comparison being run
}
\description{
provide the output of a single method for a single trial; intended primarily for troubleshooting and debugging
}

