% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{list_output_parsers}
\alias{list_output_parsers}
\title{List outputParsers}
\usage{
list_output_parsers(dsc)
}
\arguments{
\item{dsc}{the dsc to have its outputParsers listed}
}
\value{
nothing
}
\description{
List outputParsers
}
