% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list_scenarios}
\alias{list_scenarios}
\title{List all scenarios for a DSC}
\usage{
list_scenarios()
}
\arguments{
\item{scenarios}{a list of scenarios used to produce data=list(input,meta)}
}
\value{
a list of names of the scenarios
}
\description{
List all scenarios for a DSC
}

