% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{list_scenarios}
\alias{list_scenarios}
\title{List scenarios}
\usage{
list_scenarios(dsc)
}
\arguments{
\item{dsc}{the dsc to have its scenarios listed}
}
\value{
nothing
}
\description{
List scenarios
}

