% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{listoutputParsers}
\alias{listoutputParsers}
\title{List outputParsers}
\usage{
listoutputParsers(dsc)
}
\arguments{
\item{dsc}{the dsc to have its outputParsers listed}
}
\value{
nothing
}
\description{
List outputParsers
}

