% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deprecated.R
\name{listoutputParsers}
\alias{listoutputParsers}
\title{Deprecated alias to \code{\link{list_output_parsers}}.}
\usage{
listoutputParsers(...)
}
\description{
Deprecated alias to \code{\link{list_output_parsers}}.
}

