% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{listScenarios}
\alias{listScenarios}
\title{Deprecated alias to \code{\link{list_scenarios}}.}
\usage{
listScenarios(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{list_scenarios}}.}
}
\description{
Deprecated alias to \code{\link{list_scenarios}}.
}
