% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{load_example}
\alias{load_example}
\title{return the data and output for a single method for a single trial}
\usage{
load_example(dsc, seed, scenarioname, methodname)
}
\arguments{
\item{seed}{}

\item{scenario}{}

\item{method}{}
}
\value{
results list with components data, output
}
\description{
return a list containing data and output for a single method,trial
}

