% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{load_example}
\alias{load_example}
\title{return the data and output for a single method for a single trial}
\usage{
load_example(dsc, seed, scenarioname, methodname, homedir = ".")
}
\arguments{
\item{homedir}{the directory from which the dsc was run}
}
\value{
results list with components data, output
}
\description{
return a list containing data and output for a single method,trial
}
