% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{make_data}
\alias{make_data}
\title{Make all the data for a DSC}
\usage{
make_data(scenarios)
}
\arguments{
\item{scenarios}{a list of scenarios}
}
\value{
none; data are saved in files in the data subdirectory
}
\description{
Make all the data for DSC using all combinations of seed and scenario
}

