% Generated by roxygen2 (4.0.2): do not edit by hand
\name{make_data}
\alias{make_data}
\title{Make all the inputs for a DSC}
\usage{
make_data(datamaker, scenario_seedlist)
}
\arguments{
\item{datamaker}{a function for making data from seeds and scenario combinations}

\item{scenario_seedlist}{a list of integer vectors, one for each scenario. Each list gives the seeds to be used for that scenario.}
}
\value{
parameters are saved in files in the data subdirectory
}
\description{
Make all the data for DSC using all combinations of seed and scenario
}

