% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make_data}
\alias{make_data}
\title{Make all the inputs for a DSC}
\usage{
make_data(datamaker, scenario_seedlist)
}
\arguments{
\item{parammaker}{a function for making parameters from seeds and scenario combinations}

\item{seed}{(list of integers) the seeds for the pseudo-rng which identifies/indexes trials}

\item{scenario}{(list of strings) the names of the scenarios}
}
\value{
parameters are saved in files in the params subdirectory
}
\description{
Make all the parameters for DSC using all combinations of seed and scenario
}

