% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{make_data_scenario}
\alias{make_data_scenario}
\title{Make the data (inputs and meta) for a DSC for a particular scenario}
\usage{
make_data_scenario(scenario)
}
\arguments{
\item{scenario}{a list including elements fn and args, the function name for the datamaker to be used and additional arguments}
}
\value{
data are saved in files in the data subdirectory
}
\description{
Make the data (inputs and meta) for a DSC for a particular scenario
}

