% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{new.dsc}
\alias{new.dsc}
\title{Start a new (empty) dsc}
\usage{
new.dsc(name)
}
\arguments{
\item{name}{a string containing the name for your dsc}
}
\value{
dsc, an environmnet
}
\description{
returns an environment that is a dsc
}

