% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{new_dsc}
\alias{new_dsc}
\title{Start a new (empty) dsc}
\usage{
new_dsc(name, file.dir)
}
\arguments{
\item{name}{a string containing the name for your dsc}
}
\value{
dsc, an environmnet
}
\description{
returns an environment that is a dsc
}

