% Generated by roxygen2 (4.0.2): do not edit by hand
\name{reset_dsc}
\alias{reset_dsc}
\title{Removes all data, output and results for the dsc}
\usage{
reset_dsc(scenarios, methods, force = FALSE)
}
\arguments{
\item{scenarios}{a list of scenarios in the dsc}

\item{methods}{a list of methods in the dsc}

\item{force}{boolean, indicates whether to proceed without prompting user}
}
\value{
nothing; simply deletes files
}
\description{
Removes all files in results/ data/ and output/ subdirectories. Mostly useful for testing purposes.
}

