% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{reset_dsc}
\alias{reset_dsc}
\title{Removes all data, output and results for the dsc}
\usage{
reset_dsc(dsc, force = FALSE)
}
\arguments{
\item{force}{boolean, indicates whether to proceed without prompting user}

\item{scenarios}{a list of scenarios in the dsc}

\item{methods}{a list of methods in the dsc}
}
\value{
nothing; simply deletes files
}
\description{
Removes all files in scores/ meta/, input/ and output/ subdirectories. Mostly useful for testing purposes.
}

