% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{reset_dsc}
\alias{reset_dsc}
\title{Add outputfilenames to the database
#
#}
\usage{
reset_dsc(dsc, force = FALSE)
}
\arguments{
\item{dsc}{the dsc to use
#
#}

\item{force}{boolean, indicates whether to proceed without prompting user}

\item{scenarios}{a list of scenarios in the dsc}

\item{methods}{a list of methods in the dsc}
}
\value{
nothing, but modified dsc
#

nothing; simply deletes files
}
\description{
Add outputfilenames
#
}
\details{
#
}

