% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{reset_method}
\alias{reset_method}
\title{Removes all output and results for a method}
\usage{
reset_method(dsc, methodname, force = FALSE)
}
\arguments{
\item{force}{boolean, indicates whether to proceed without prompting user (prompt is to be implemented)}

\item{method}{string indicating name of methods to remove output}
}
\value{
nothing; simply delets files
}
\description{
Removes all output and results for a method; primary intended purpose is to force re-running of that method.
Works only for unix look-alikes?
}

