% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_dsc}
\alias{run_dsc}
\title{Run all methods on all scenarios for a DSC.}
\usage{
run_dsc(dsc, scenariosubset = NULL, methodsubset = NULL,
  seedsubset = NULL)
}
\arguments{
\item{scenariosubset}{A vector of the names of the scenarios to
actually make and run.}

\item{methodsubset}{A vector of the names of the methods to run
(default is to run all of them).}

\item{seedsubset}{A vector of which seeds to run. e.g., \code{1:2}
would do seeds 1 and 2 (for scenarios that use that seed).}

\item{scenarios}{A list of scenarios used to produce
\code{data = list(input,meta)}.}

\item{methods}{A list of methods to turn \code{data$input} into output.}

\item{scorefn}{A function that takes output and scores it against data.}
}
\value{
Data frame of results from all methods run on all scenarios.
}
\description{
Run all methods on all scenarios for a DSC.
}
