% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{runParser}
\alias{runParser}
\title{Run a parser in a dsc}
\usage{
runParser(dsc, parsername)
}
\arguments{
\item{dsc}{the dsc to use}

\item{parsername}{the name of the parser function to be run}
}
\value{
nothing, but outputs files to output/ directories
}
\description{
Run the named parser to convert one type of output to another type
}

