% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{score_method_singletrial}
\alias{score_method_singletrial}
\title{Score a method on a single trial and save results}
\usage{
score_method_singletrial(dsc, seed, scenario, method, score)
}
\arguments{
\item{seed}{the seed to score}

\item{scenario}{the scenario to score}

\item{method}{the method to score}

\item{scorefn}{a function that scores output based on comparisons with input, parameters and metadata}
}
\value{
results, a list of appropriate format to be determined by the comparison being run (maybe required to be a dataframe?)
}
\description{
Score results of a single method for a single trial and produce (and save) corresponding results
}

