% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{sourceDir}
\alias{sourceDir}
\title{Deprecated alias to \code{\link{source_dir}}.}
\usage{
sourceDir(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{source_dir}}.}
}
\description{
Deprecated alias to \code{\link{source_dir}}.
}
